#210: change regoff_t from int to ptrdiff_t

+++ src/onigposix.h
@@ -74,7 +74,7 @@ extern "C" {
 #define REG_POSIX_ENCODING_UTF16_LE  5
 
 
-typedef int onig_posix_regoff_t;
+typedef ptrdiff_t  onig_posix_regoff_t;
 
 typedef struct {
   onig_posix_regoff_t  rm_so;
